+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and includes an additional section for migration notes.
 - *ORCA-256* Added AWS API Gateway in lambdas.tf for the catalog reporting lambda.
 - *ORCA-227* Added modules/secretsmanager directory that contains terraform code for deploying AWS secretsmanager.
 - *ORCA-177* Added AWS API Gateway in lambdas.tf for the request_status_for_granule and request_status_for_job lambdas.
+- *ORCA-257* orca_catalog_reporting lambda now returns data from actual catalog.
 
 ### Changed
 - *ORCA-287* Updated extract_filepaths_for_granule to [new Cumulus file format](https://github.com/nasa/cumulus/blob/master/packages/schemas/files.schema.json).